Move TopicViewModelFactory from MainActivity.kt to TopicViewModel.kt
