+++ MemTopic/app/src/main/java/com/nolbee/memtopic/database/TopicViewModel.kt
@@ -1,6 +1,7 @@
 package com.nolbee.memtopic.database
 
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -24,5 +25,16 @@ class TopicViewModel(private val repository: TopicRepository) : ViewModel() {
         viewModelScope.launch {
             repository.deleteTopic(topic)
         }
+    }
+}
+
+class TopicViewModelFactory(private val repository: TopicRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(TopicViewModel::class.java)) {
+            val viewModel = TopicViewModel(repository)
+            return modelClass.cast(viewModel)
+                ?: throw IllegalArgumentException("Cannot cast to ConfigViewModel")
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
     }
 }
